Level design and research
Update was later than normal due to a combination of a Dental apointment that was followed by an internet outage.

Created main teleporter room that gets the player to "floor 2". In order to activate this I will have a 2 part puzzle where the last part activates it.

Platforming is easy enough to create, the question is how do I justify when I prevent the player from crossing into one area or not? Walls by default, but until I fix the lighting issue there are pretend walls.

Level design is much slower than anticipated, and will be gradualy done. There is also the realization for the need for lighting which does not seem present.