Props placed for puzzle 1 floor 1, puzzle 2 floor 1 in progress
Enemy generators sill have to be placed, nav mesh to be made, but the first puzzle area is nearly done. What it will do is not quite determined, maybe open the way to puzzle 2.

Main reason for nav mesh not being made is because this first floor is still being developed, and I don't want an area enemies can't cross because they have no idea it's there.