Added room 2
Had to change a few thing up, when I discovered the innitial archway I had planned to use as the enterance to the second puzzle room had turned out not to be a solid object pr have a proper mesh for making it solid. Replaced it with a gate that will eventualy open when the switch is pulled, and a stairway for the drop.

Playtesting discovered hidden area, will be filled in a later date with rewards for finding it.

Lighting still an issue

Next steps are to create the "second" floor, which is much less complex than the first,
Nav mesh baking
Enemy genorator placement
Enemy genorator upgrade (on entering an area activate)
Lighting

Optional hopefuls:
Behavioral changes in creatures
More unique traps than just one supprise one in the entire map (posibly loading the second room with "Fake" presure plates to start the parkour puzzle)